Attach to active clients on setup
This improves the lazy loading capabilities of this plugin. If the user
has enabled `auto_attach`, during the setup the plugin will attach
itself to all already active clients. This is an addition to the already
existing auto-command that will attach to new clients.
This allows to source/load the plugin any time and it will work right
away having the contact to all clients.